scripts/AsciiDocBasics.gradle: fulfill different plantUML path requirements by backends
The commit

    97df9b9 scripts/AsciiDocBasics.gradle: fix inline plantuml referencing

introduces same path for all backends to save generated images from inline
plantUML code.  Unfortunately, it introduces a bug for Concluence backends
and amplifies a bug for the HTML5 backend.

This commit moves the plantUMLDir back to html5 and pdf backend, which should
avoid harming Confluence backend at all. Further, the value for the html5
backend is computed to a relative path for deploying the entire directory
structure to various locations.

https://github.com/docToolchain/docToolchain/issues/582 should be fixed by
this change.

Signed-off-by: Jens Rehsack <sno@netbsd.org>