remove invalid theme folder before throwing GradleException
As specified in https://github.com/docToolchain/docToolchain/issues/658, subsequent runs of generateSite wih a missing or invalid DTC_SITETHEME will not result in build failure as expected.

The solution that fixes the issue for me was to swap lines 183 and 184 and remove the themeFolder before throwing the exception. 

No further analysis of the generateSite task done.