Jody's edits
Changed git to GitHub but if you actually meant Git, then use capital G.

I have used title case for the 'Improve This Doc' and 'Create an Issue' button labels following the Chicago Manual of Style capitalisation conventions. If these labels are actually rendered as 'improve this doc' and 'create-an-issue' when this task is executed then please change back.