Clarify that ng branch is the future in README
The README in the ng branch said "master is the right thing" while
the README in the master branch said "ng is the right thing".
From the discussion in issue #716 I conclude that master is deprecated
and ng is the branch to go (supported by the fact that ng is default).
See also: https://github.com/docToolchain/docToolchain/issues/716#issuecomment-993917342