Update asciidoc2confluence.groovy
Fix a problem with different compare results of "pageExists" between docToolchain and Confluence.

If the title of a page has trailing or leading spaces it will be removed by Confluence on upload. This leads to different results for the comparison in docToolchain == "No" and Confluence == "Yes".

This in turn triggers an upload of the page to Confluence - where it is rejected.
Finally the user gets a
"A page with this title already exists: A page already exists with the title <TITLE> in the space with key <SPACE>
please check the ancestorId in your config file"