Fixes using the wrong `arch` on macOS (Darwin)
If you happen to have another `arch` binary or script on ${PATH}
this was used instead of /usr/bin/arch - with disturbing results.
The patch ensures the right executable is used to perform
docToolchain on macOS.

For shell scripts it is always a good idea to use explicit paths
to become independent of arbitrary commands on the ${PATH}).